Step 3b (drift fix): move UseCaseAttributes from model/ to api/
Per DIR-PACKAGE-DRIFT-FIX-punit step 3b. UseCaseAttributes is the
author-side carrier of attributes (warmup, maxConcurrent, …) declared
on a use case; it belongs on the public api root, not in the
internal model/ catch-all.

Touched:
- git mv main + test from model/ to api/.
- Rewrite package declarations.
- Update import paths across all consumers in punit-core,
  punit-report, punit-sentinel (single bulk sed; ~14 files).

Architecture-test bookkeeping: noTopLevelModelPackage rule's
freeze-store shrunk from 10 → 9 entries.

The punitexamples repo had no direct UseCaseAttributes import
(consumes via punit-junit5 / api conduits only); composite-build
green.
