+++ punit-core/archunit_store/90083594-5684-45be-a3cf-87dc0d2f9509
@@ -7,4 +7,3 @@ Class <org.javai.punit.model.ExpirationStatus$NoExpiration> does reside in a pac
 Class <org.javai.punit.model.ExpirationStatus$Valid> does reside in a package 'org.javai.punit.model..' in (ExpirationStatus.java:0)
 Class <org.javai.punit.model.ExpirationStatus> does reside in a package 'org.javai.punit.model..' in (ExpirationStatus.java:0)
 Class <org.javai.punit.model.TerminationReason> does reside in a package 'org.javai.punit.model..' in (TerminationReason.java:0)
-Class <org.javai.punit.model.UseCaseAttributes> does reside in a package 'org.javai.punit.model..' in (UseCaseAttributes.java:0)
