+++ punit-core/src/test/java/org/javai/punit/api/UseCaseAttributesTest.java
@@ -1,4 +1,4 @@
-package org.javai.punit.model;
+package org.javai.punit.api;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
