+++ punit-sentinel/src/test/java/org/javai/punit/sentinel/testsubjects/SentinelTestSubjects.java
@@ -10,7 +10,7 @@ import org.javai.punit.api.ProbabilisticTest;
 import org.javai.punit.api.TestIntent;
 import org.javai.punit.controls.budget.CostBudgetMonitor.TokenMode;
 import org.javai.punit.model.TerminationReason;
-import org.javai.punit.model.UseCaseAttributes;
+import org.javai.punit.api.UseCaseAttributes;
 import org.javai.punit.verdict.ProbabilisticTestVerdict;
 import org.javai.punit.verdict.PUnitVerdict;
 import org.javai.punit.verdict.VerdictSinkBus;
